# src/doc/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Edugram API",
    "description": "Edugram Backend API",
    "version": "1.0.0"
  },
  "host": "katowice.edugram.io",
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "paths": {
...
    "/api/education/courses/{courseId}/reviews": {
      "get": {
        "tags": [
          "Courses"
        ],
        "description": "",
        "parameters": [
          {
            "name": "courseId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Courses"
        ],
        "description": "",
        "parameters": [
          {
            "name": "courseId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "example": "any"
                },
                "review": {
                  "example": "any"
                },
                "rating": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/education/courses/{courseId}/reviews/{reviewId}": {
      "patch": {
        "tags": [
          "Courses"
        ],
        "description": "",
        "parameters": [
          {
            "name": "courseId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "reviewId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "example": "any"
                },
                "review": {
                  "example": "any"
                },
                "rating": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Courses"
        ],
        "description": "",
        "parameters": [
          {
            "name": "courseId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "reviewId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/education/courses/{courseId}/review-record": {
      "get": {
        "tags": [
          "Courses"
        ],
        "description": "",
        "parameters": [
          {
            "name": "courseId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/education/courses/{courseId}/review-records": {
      "get": {
        "tags": [
          "Courses"
        ],
        "description": "",
        "parameters": [
          {
            "name": "courseId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/education/courses/{courseId}/{courseVersion}/review-record": {
      "get": {
        "tags": [
          "Courses"
        ],
        "description": "",
        "parameters": [
          {
            "name": "courseId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "courseVersion",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Courses"
        ],
        "description": "",
        "parameters": [
          {
            "name": "courseId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "courseVersion",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/review-records/{reviewRecordId}": {
      "get": {
        "tags": [
          "Review Records"
        ],
        "description": "",
        "parameters": [
          {
            "name": "reviewRecordId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/review-records/{reviewRecordId}/start-review": {
      "post": {
        "tags": [
          "Review Records"
        ],
        "description": "",
        "parameters": [
          {
            "name": "reviewRecordId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "adminId": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/review-records/{reviewRecordId}/approve": {
      "post": {
        "tags": [
          "Review Records"
        ],
        "description": "",
        "parameters": [
          {
            "name": "reviewRecordId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "adminId": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/review-records/{reviewRecordId}/reject": {
      "post": {
        "tags": [
          "Review Records"
        ],
        "description": "",
        "parameters": [
          {
            "name": "reviewRecordId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "rejectionReason": {
                  "example": "any"
                },
                "adminId": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
...
}